@@ -1,7 +1,7 @@
 import '../../utils/uni_enums.dart';
 
 class UniPayAddress {
-  ///* Customer address name e.g: Olaya street, Al qadir
+  ///* Customer address name e.g: Olaya street, Al Ghadir
   late String addressName;
 
   ///* Address city
@@ -10,16 +10,21 @@ class UniPayAddress {
   ///* Customer country
   late UniPayCountry country;
 
+  ///* Address zip code default is `12211 -> Riyadh`
+  late String zipCode;
+
   UniPayAddress({
     required this.addressName,
     required this.city,
     this.country = UniPayCountry.sa,
+    this.zipCode = "12211",
   });
 
   UniPayAddress.fromJson(Map<String, dynamic> data) {
     addressName = data['address_name'];
     city = data['city'];
     country = UniPayCountry.values[data['country']];
+    zipCode = data['zip_code'];
   }
 
   Map<String, dynamic> toJson() {
@@ -27,6 +32,7 @@ class UniPayAddress {
     data['address_name'] = addressName;
     data['city'] = city;
     data['country'] = country.index;
+    data['zip_code'] = zipCode;
     return data;
   }
 }
